csharp telemetry generator (#35)
This is a C# Telemetry Generator for use with the AWS Toolkit for Visual Studio

See telemetry/csharp/ToolkitTelemetryGenerator.Tests/expectedCode.txt for an example of generated code.

When integrated into an application, the autogenerated code expects ITelemetryLogger to be implemented. The application code then records a metric using extension methods on ITelemetryLogger.