Proof of Concept: Reusable JSON and Yaml servers powered by JSON Schemas (#497)
This is part of a series of changes that establishes an LSP monorepo that we can perform experiments on.

This change contains proof of concept implementations of JSON (`JsonSchemaServer`) and Yaml (`YamlSchemaServer`) language servers, that operate on a JSON Schema. The classes can be instantiated with a url to a JSON Schema, and it will provide completion/validation based on that schema. A follow-up change will illustrate how these are leveraged. These servers are not optimized (or de-duplicated) at this time.

To support the proof of concept, code has been added to the core library. The `AwsLanguageService` interface is a concept that will be explored further as we grow out the monorepo. This may provide a boundary that contains a document's (or language)'s domain knowledge, without coupling to the LSP Connection.