Buildspec servers (#498)
This is part of a series of changes that establishes an LSP monorepo that we can perform experiments on.

This change demonstrates how we can create a language server around a CodeBuild Buildspec JSON Schema, building on the proof of concept reusable components defined in https://github.com/aws/aws-toolkit-common/pull/497 .

Three separate language servers have been created:
- one that applies buildspec logic to JSON files (BuildspecJsonServer)
- one that applies buildspec logic to Yaml files (BuildspecYamlServer)
- one that applies buildspec logic to both JSON and Yaml files (BuildspecServer)

BuildspecJsonServer and BuildspecYamlServer shows how we can leverage something that just knows how to wrap a JSON Schema. We don't know yet if we'd want to host disparate language servers like these in the IDE, which is where BuildspecServer comes in. We don't necessarily want to compose language servers together, because there will be collisions at the lsp connection callback level. BuildspecServer illustrates how one level lower (language services) may be a good reusability layer.

To support this change, MutuallyExclusiveLanguageService (naming is hard) shows how we can compose AwsLanguageService implementations in order to produce something that knows how to handle JSON and YAML files.

In a follow up change, I will restore the sample language clients, which will allow us to run and debug this language server while developing in the repo.