Buildspec binary; Relocate sample language clients (#500)
This is part of a series of changes that establishes an LSP monorepo that we can perform experiments on.

This change demonstrates how we can create a standalone language server binary that can be integrated into IDEs. A binary is created using the Buildspec server created in https://github.com/aws/aws-toolkit-common/pull/498.

This change also relocates the sample language client extensions to their new location in the monorepo. The sample VS Code client has been adjusted to run using its updated pathing, and a launch profile has been set up that is capable of running the buildspec server.