Clear security scan findings on workspace/didChangeWorkspaceFolders event (#207)
This change fixes a bug where when a project was unloaded in the IDE and a new one opened, the security scan findings(diagnostics) associated with the previous project did not get cleared.
This change adds handling for the workspace/didChangeWorkspaceFolders lsp notification sent by the client. The handler clears diagnostics associated with the previous security scan run(project).