fix(amazonq): add regionalization support to .NET Transform server (#952)
This change updates the .NET Transform server to use AmazonQTokenServiceManager when producing the service client. This way, the service calls are able to provide the developer profile arn or region details for IdC based connections.

This change is based on #859, which made a similar change for security scanning.

I worked on this change with @shruti0085 , and we tested it by using the Visual Studio Toolkit to perform a .NET Transform using a local build of the server. We were able to see the profileArn field being passed along with all of the .NET Transform related service calls.

License
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.