fix(amazonq): avoid ERR_UNSUPPORTED_ESM_URL_SCHEME error when loading indexer on Windows (#1135)
## Problem

The codewhisperer server fails to load the project context indexer on windows, raising a `ERR_UNSUPPORTED_ESM_URL_SCHEME ` failure. This happens when the library is imported with a regular path on Windows:

https://github.com/aws/language-servers/blob/e4d4ef026c8a60cc1ddf08c981340a902d628016/server/aws-lsp-codewhisperer/src/shared/localProjectContextController.ts#L142-L143

Some information from https://github.com/nodejs/node/issues/31710 indicates that code running on Windows must use URLs so that local file paths (including drive letters) are not mistakenly interpreted as things like http urls.

## Solution

Wrap the path in a URL when running on Windows.