fix: bubble up throttling error for ListAvailableProfiles API to client (#1127)
Problem
When a throttling exception is thrown with the ListAvailableProfiles API, the error is swallowed and client is informed of a generic failure without information on the cause of the failure. Since clients do not have this information, a vague message is shown to the user similar to Unable to retrieve profiles. We want this to be bubbled upstream to the client

Solution
Whenever a throttling exception is thrown with the ListAvailableProfiles API, a ResponseError with a specific awsErrorCode E_AMAZON_Q_PROFILE_THROTTLING is now sent back for the client to interpret and use for showing an appropriate message to the user.