fix: add support for determing workspace folder with root uri/path on initialize (#1210)
LSP lifecycle for VS extension is controlled by Visual Studio. It currently does not set the workspaceFolder as part of the Initialize Params of the LSP initialize request, however it does send this information via rootUri and rootPath properties. The workspaceFolders is utilized for the local project context indexing and so in order to support that, this change extends logic to determine the workspaceFolder to use the root properties when workspaceFolders is not set.