fix: wrap load chats on ready in try-catch (#1289)
Problem
Currently when the chat loads up a UI ready event is sent to the lsp, to load up previous chats and start indexing. If loading of previous chats fail, the LSP process crashes. This is particularly evident when openTab event has not been handled by the client.
An error in that flow should not crash the LSP process or block the rest of the indexing from continuing.

Solution
This change surrounds the load chat logic within a try catch of its own to allow the rest of the indexing logic to proceed and not crash the server process