fix: use updated version of vecLib and use local context controller to raise context command updates (#1479)
## Problem
`DidWorkspaceFolderChange` events did not update context commands. This issue is particularly pronounced in Visual Studio where this event is used to send information about projects contained in a solution on intial load which should be indexed.
In addition, the updateIndex calls for vecLib was gated by the indexingEnabled flag which meant when a change event came with the indexer off, the context commands were not updated. The context command updates should happen unconditionally even when indexer is disabled.
## Solution
This change addresses the issues above:
* A new version of the vecLib has been added to the local development copy with the tgz file. A corresponding release PR is being tracked in a separate repo.
* The update above adds support for updating indexer and context commands when workspace folders change.
* A corresponding change has been made to the language server to support sending the workspace folders updates to the vecLib update index call.
* The context command provider and controller were very strongly coupled causing issues updating the two independently. With this change, the controller now becomes the sole authority that raises an event to context command provider to get and push latest context commands after indexing finishes.
* IndexingEnabled flag has been removed as check from the updateIndex calls. It is now expected that the vecLib knows when to not index locally(for @workspace) when the buildIndex('default') is invoked with this value.