chore: add a test LFS file to the repo (#2013)
This change is a test using LFS. It has one small hello world zip (containing a text file that says "hello world").

The purpose of this change is to test some internal processes to ensure they are able to successfully consume the repo once it has LFS contents in it.

* chore: define LFS content in a binaries folder as a test
* chore: add sample LFS file for a test