feat: nep auto trigger (#2424)
## Problem
The existing edit trigger criteria is too loose (1) non empty right context (2) edit history in the past 20s.


## Solution
Build a logistic regression classifier and use users editor state as the features of the formula.
<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots if applicable
-->

## License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.